refactor: decompose Parley into ConnectionManager and SendPipeline
Internal-only move refactor; the public API and index.ts exports are
unchanged and all 861 unit/integration tests plus the 5 Playwright E2E
tests pass without modification.

- ConnectionManager (new, 462 lines): channel ownership and creation,
  connect/disconnect lifecycle, heartbeat failure handling, internal
  system-message handlers, connection state events
- SendPipeline (new, 576 lines): send/broadcast, pending-request map
  and response correlation, timeout/retry, payload size validation,
  rate limiting
- Parley.ts (1544 -> 696 lines): facade with the public API, message
  receiving/request handling, analytics wiring, heartbeat ownership
- Collaborators receive narrow callbacks instead of whole objects to
  avoid circular references
